requirements.txt
pyyaml==6.0.1
click==8.1.7
jinja2==3.1.4